完善获取卡片和保存卡片接口

+++ miniAPP-backend/miniAPP-mapper/src/main/java/com/miniAPP/mapper/FrCardMapper.java
@@ -45,5 +45,5 @@ public interface FrCardMapper extends MyMapper<FrCard> {
      *
      * @param labelID
      */
-    List<FrCard> queryAllCardByLabelID(Long labelID);
+    List<FrCard> queryAllCardByLabelID(Integer labelID);
 }
